Reverted back hooks and context
